+++ ci/docker-tag-remote-version-as-latest.sh
@@ -1,8 +1,6 @@
 #!/bin/bash
 # Tag a remote version as latest
 set -euo pipefail
-export DOCKER_CLI_EXPERIMENTAL=enabled
-export DOCKER_BUILDKIT=1
 REPO=docker.io/ajslater/codex
 VERSION=$1