+++ ci/versions-env-filename.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+# Set the env filename var
+ARCH=$(./ci/machine-arch.sh)
+export ENV_FN=./.env-${ARCH}