+++ docker/docker-install-api-deps.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-# install dependencies for hitting the hub.docker.com api directly
-set -euxo pipefail
-
-sudo apt-get update
-sudo apt-get install curl jq