+++ NEWS.md
@@ -6,9 +6,18 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.8.17
+
+- Features
+    - Python 3.14 support.
+
+- Fixes
+    - Fix for codex sometimes shutting when asked to restart.
+
 ## v1.8.16
 
 - Don't support Python 3.14 until dependencies do.
+- Docker image is now based on trixie debian
 
 ## v1.8.15